refactor(eth): extract APDU dispatch and app context from main.c
