+++ src/main.c
@@ -83,48 +83,6 @@ const internalStorage_t N_storage_real;
 const caller_app_t *g_caller_app = NULL;
 const chain_config_t *g_chain_config;
 
-void reset_app_context(void) {
-    if (appState == APP_STATE_SIGNING_MESSAGE) {
-        message_cleanup();
-    }
-    eip712_v1_context_deinit();
-    appState = APP_STATE_IDLE;
-    G_called_from_swap = false;
-    G_swap_response_ready = false;
-    G_swap_checked = false;
-    pluginType = PLUGIN_TYPE_NONE;
-#ifdef HAVE_ETH2
-    eth2WithdrawalIndex = 0;
-#endif
-    memset((uint8_t *) &tmpCtx, 0, sizeof(tmpCtx));
-    forget_known_assets();
-    if (txContext.store_calldata) {
-        gcs_cleanup();
-    }
-    trusted_name_cleanup();
-    enum_value_cleanup();
-    map_entry_cleanup();
-    // Release the tx-signing keccak context.
-    if (g_tx_hash_ctx != NULL) {
-        APP_MEM_FREE_AND_NULL((void **) &g_tx_hash_ctx);
-    }
-    memset((uint8_t *) &txContext, 0, sizeof(txContext));
-    memset((uint8_t *) &tmpContent, 0, sizeof(tmpContent));
-    clear_safe_account();
-    ui_all_cleanup();
-    proxy_cleanup();
-    clear_gating();
-}
-
-void app_quit(void) {
-    network_info_cleanup(NULL);
-#ifdef HAVE_ADDRESS_BOOK
-    address_book_contact_cleanup();
-#endif  // HAVE_ADDRESS_BOOK
-    reset_app_context();
-    app_exit();
-}
-
 uint16_t io_seproxyhal_send_status(uint16_t sw, uint32_t tx, bool reset, bool idle) {
     if (reset) {
         reset_app_context();
@@ -135,193 +93,6 @@ uint16_t io_seproxyhal_send_status(uint16_t sw, uint32_t tx, bool reset, bool id
         ui_idle();
     }
     return (uint16_t) (err < 0 ? 1 : 0);
-}
-
-const uint8_t *parseBip32(const uint8_t *dataBuffer, uint8_t *dataLength, bip32_path_t *bip32) {
-    if (*dataLength < 1) {
-        PRINTF("Invalid data\n");
-        return NULL;
-    }
-
-    bip32->length = *dataBuffer;
-
-    dataBuffer++;
-    (*dataLength)--;
-
-    if (*dataLength < sizeof(uint32_t) * (bip32->length)) {
-        PRINTF("Invalid data\n");
-        return NULL;
-    }
-
-    if (bip32_path_read(dataBuffer, (size_t) dataLength, bip32->path, (size_t) bip32->length) ==
-        false) {
-        PRINTF("Invalid Path data\n");
-        return NULL;
-    }
-    dataBuffer += bip32->length * sizeof(uint32_t);
-    *dataLength -= bip32->length * sizeof(uint32_t);
-
-    return dataBuffer;
-}
-
-/**
- * Main APDU dispatcher.
- *
- * This function validates the APDU class byte and routes each supported
- * instruction (INS) to its dedicated handler. Handlers are responsible for
- * parsing command data, producing response bytes, and updating `*tx` with the
- * response length when applicable. The returned status word (`sw`) is sent
- * back to the host by the caller.
- *
- * Maintenance note:
- * - Keep one case per INS value.
- * - Prefer delegating parsing/business logic to dedicated handlers.
- * - Preserve existing state-reset and cleanup behavior when adding new cases.
- */
-static uint16_t handleApdu(command_t *cmd, uint32_t *tx) {
-    uint16_t sw = SWO_NO_RESPONSE;
-
-    if (cmd->cla != CLA) {
-        return SWO_INVALID_CLA;
-    }
-
-    switch (cmd->ins) {
-        // Key derivation and app configuration
-        case INS_GET_PUBLIC_KEY:
-            forget_known_assets();
-            sw = handle_get_public_key(cmd->p1, cmd->p2, cmd->data, cmd->lc, tx);
-            break;
-
-        // Provide off-chain metadata used to enrich the signing UI
-        case INS_PROVIDE_ERC20_TOKEN_INFORMATION:
-            sw = handle_provide_erc20_token_information(cmd->p1, cmd->p2, cmd->lc, cmd->data, tx);
-            break;
-
-        case INS_PROVIDE_NFT_INFORMATION:
-            sw = handle_provide_nft_information(cmd->p1, cmd->p2, cmd->lc, cmd->data, tx);
-            break;
-
-        // External / internal plugin selection
-        case INS_SET_EXTERNAL_PLUGIN:
-            sw = handle_set_external_plugin(cmd->data, cmd->lc);
-            break;
-
-        case INS_SET_PLUGIN:
-            sw = handle_set_plugin(cmd->data, cmd->lc);
-            break;
-
-        case INS_PERFORM_PRIVACY_OPERATION:
-            sw = handle_perform_privacy_operation(cmd->p1, cmd->p2, cmd->data, cmd->lc, tx);
-            break;
-
-        // Transaction and message signing
-        case INS_SIGN:
-            sw = handle_sign(cmd->p1, cmd->p2, cmd->data, cmd->lc);
-            break;
-
-        case INS_GET_APP_CONFIGURATION:
-            sw = handle_get_app_configuration(tx);
-            break;
-
-        case INS_SIGN_PERSONAL_MESSAGE:
-            forget_known_assets();
-            sw = handle_sign_personal_message(cmd->p1, cmd->data, cmd->lc);
-            break;
-
-        case INS_SIGN_EIP_712_MESSAGE:
-            switch (cmd->p2) {
-                case P2_EIP712_V0_IMPLEM:
-                    forget_known_assets();
-                    sw = handle_sign_eip712_message_v0(cmd->p1, cmd->data, cmd->lc);
-                    break;
-                case P2_EIP712_V1_IMPLEM:
-                    sw = handle_eip712_v1_sign(cmd->data, cmd->lc);
-                    break;
-                default:
-                    sw = SWO_WRONG_P1_P2;
-            }
-            break;
-
-#ifdef HAVE_ETH2
-        case INS_GET_ETH2_PUBLIC_KEY:
-            forget_known_assets();
-            sw = handle_get_eth2_public_key(cmd->p1, cmd->p2, cmd->data, cmd->lc, tx);
-            break;
-
-        case INS_SET_ETH2_WITHDRAWAL_INDEX:
-            sw = handle_set_eth2_withdrawal_index(cmd->p1, cmd->p2, cmd->data, cmd->lc);
-            break;
-#endif  // HAVE_ETH2
-
-        // EIP-712 full implementation: structure definition, instantiation, filtering
-        case INS_EIP712_STRUCT_DEF:
-            sw = handle_eip712_v1_struct_def(cmd->p2, cmd->data, cmd->lc);
-            break;
-
-        case INS_EIP712_STRUCT_IMPL:
-            sw = handle_eip712_v1_struct_impl(cmd->p1, cmd->p2, cmd->data, cmd->lc);
-            break;
-
-        case INS_EIP712_FILTERING:
-            sw = handle_eip712_v1_filtering(cmd->p1, cmd->p2, cmd->data, cmd->lc);
-            break;
-
-        // Generic clear-signing protocol (GTP) and trusted data provisioning
-        case INS_GET_CHALLENGE:
-            sw = handle_get_challenge(tx);
-            break;
-
-        case INS_PROVIDE_TRUSTED_NAME:
-            sw = handle_trusted_name(cmd->p1, cmd->data, cmd->lc);
-            break;
-
-        case INS_PROVIDE_ENUM_VALUE:
-            sw = handle_enum_value(cmd->p1, cmd->p2, cmd->lc, cmd->data);
-            break;
-
-        case INS_GTP_TRANSACTION_INFO:
-            sw = handle_tx_info(cmd->p1, cmd->p2, cmd->lc, cmd->data);
-            break;
-
-        case INS_GTP_FIELD:
-            sw = handle_field(cmd->p1, cmd->p2, cmd->lc, cmd->data);
-            break;
-
-        case INS_PROVIDE_PROXY_INFO:
-            sw = handle_proxy_info(cmd->p1, cmd->p2, cmd->lc, cmd->data);
-            break;
-
-        case INS_PROVIDE_NETWORK_CONFIGURATION:
-            sw = handle_network_info(cmd->p1, cmd->p2, cmd->data, cmd->lc, tx);
-            break;
-
-#ifdef HAVE_TRANSACTION_CHECKS
-        case INS_PROVIDE_TX_SIMULATION:
-            sw = handle_tx_simulation(cmd->p1, cmd->p2, cmd->data, cmd->lc);
-            break;
-#endif
-
-        case INS_SIGN_EIP7702_AUTHORIZATION:
-            sw = handle_sign_eip7702_authorization(cmd->p1, cmd->data, cmd->lc);
-            break;
-
-        case INS_PROVIDE_SAFE_ACCOUNT:
-            sw = handle_safe_account(cmd->p1, cmd->p2, cmd->data, cmd->lc);
-            break;
-
-        case INS_PROVIDE_GATING:
-            sw = handle_gating(cmd->p1, cmd->p2, cmd->data, cmd->lc);
-            break;
-
-        case INS_PROVIDE_MAP_ENTRY:
-            sw = handle_map_entry(cmd->p1, cmd->p2, cmd->lc, cmd->data);
-            break;
-
-        default:
-            sw = SWO_INVALID_INS;
-            break;
-    }
-    return sw;
 }
 
 void app_main(void) {
